extract logic into new Utilities class for reuse
