+++ core/domain/services/graphql/connection_resolvers/AbstractConnectionResolver.php
@@ -3,8 +3,10 @@
 namespace EventEspresso\core\domain\services\graphql\connection_resolvers;
 
 use EE_Base_Class;
+use EventEspresso\core\domain\services\graphql\Utilities;
+use EventEspresso\core\services\loaders\LoaderFactory;
+use Exception;
 use WPGraphQL\Data\Connection\AbstractConnectionResolver as WPGraphQLConnectionResolver;
-use GraphQLRelay\Relay;
 
 /**
  * Class AbstractConnectionResolver
@@ -17,6 +19,23 @@ use GraphQLRelay\Relay;
 abstract class AbstractConnectionResolver extends WPGraphQLConnectionResolver
 {
 
+
+    /**
+     * @var Utilities
+     */
+    private $utilities;
+
+
+    /**
+     * @return Utilities
+     */
+    public function getUtilities()
+    {
+        if (! $this->utilities instanceof Utilities) {
+            $this->utilities = LoaderFactory::getLoader()->getShared(Utilities::class);
+        }
+        return $this->utilities;
+    }
 
     /**
      * Determine whether the Query should execute. If it's determined that the query should
@@ -56,7 +75,7 @@ abstract class AbstractConnectionResolver extends WPGraphQLConnectionResolver
      * This checks the $args to determine the amount requested
      *
      * @return int|null
-     * @throws \Exception
+     * @throws Exception
      */
     // phpcs:ignore PSR1.Methods.CamelCapsMethodName.NotCamelCaps
     public function get_amount_requested()
@@ -145,55 +164,8 @@ abstract class AbstractConnectionResolver extends WPGraphQLConnectionResolver
      * @param array $id_fields   The fields to convert from global IDs to DB IDs.
      * @return array
      */
-    protected function sanitizeWhereArgsForInputFields(
-        array $where_args,
-        array $arg_mapping,
-        array $id_fields
-    ) {
-        $query_args = [];
-
-        foreach ($where_args as $arg => $value) {
-            if (! array_key_exists($arg, $arg_mapping)) {
-                continue;
-            }
-
-            if (is_array($value) && ! empty($value)) {
-                $value = array_map(
-                    static function ($value) {
-                        if (is_string($value)) {
-                            $value = sanitize_text_field($value);
-                        }
-                        return $value;
-                    },
-                    $value
-                );
-            } elseif (is_string($value)) {
-                $value = sanitize_text_field($value);
-            }
-            $query_args[ $arg_mapping[ $arg ] ] = in_array($arg, $id_fields, true)
-                ? $this->convertGlobalId($value)
-                : $value;
-        }
-
-        /**
-         * Return the Query Args
-         */
+    protected function sanitizeWhereArgsForInputFields(array $where_args, array $arg_mapping, array $id_fields) {
+        $query_args = $this->getUtilities()->sanitizeWhereArgs($where_args, $arg_mapping, $id_fields);
         return ! empty($query_args) && is_array($query_args) ? $query_args : [];
-    }
-
-
-    /**
-     * Converts global ID to DB ID.
-     *
-     * @param string|string[] $ID
-     * @return mixed
-     */
-    protected function convertGlobalId($ID)
-    {
-        if (is_array($ID)) {
-            return array_map([ $this, 'convertGlobalId' ], $ID);
-        }
-        $parts = Relay::fromGlobalId($ID);
-        return ! empty($parts['id']) ? $parts['id'] : null;
     }
 }
